tests/integration: separate crypto tests
